Render templates in a separate class
This allows the ability to render templates outside of the Govspeak core
class, which will be useful to render templates in post-processing.
